Fix all SonarQube reliability complaints across modules
- BenchmarkImporter: re-interrupt thread on InterruptedException
- Binpacking: refactor constructor chain so minBinNumber is set in
  primary constructor, eliminating field override after this() call
- CumulativeUnary: refactor constructor chain so doProfile/doEdgeFind
  are set in primary constructor, eliminating field override after this()
- CumulativePrimary, TupleUtils, GlobalConstraints: override
  equals/hashCode/toString in records with array components to use
  Arrays.equals/deepEquals instead of reference equality
- Regular: remove dead conditional (both branches returned 0)
- RegularExpressionParser: avoid reusing method parameters as locals
- CeilPeqX/FloorPeqX: cast int operand to double before
  arithmetic to prevent potential integer overflow
- FloatIntervalDomain: rename private cloneAndInstall to
  cloneAndInstallFloat to avoid shadowing parent class method

+++ jacop-floats/src/main/java/org/jacop/floats/constraints/CeilPeqX.java
@@ -60,7 +60,7 @@ public class CeilPeqX extends AbstractRoundPeqX {
 
     do {
       // ceil(p) = x, n - 1 < x <= n
-      p.domain.in(store.level, p, Math.nextUp(x.min() - 1), x.max());
+      p.domain.in(store.level, p, Math.nextUp((double) x.min() - 1), x.max());
 
       store.propagationHasOccurred = false;
